# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: WSO2 Web App Hosting BFF API
  version: 1.0.0
  description: |
    Backend-for-frontend of the Web App Hosting console. The console's
    `frontend/src/api/*.ts` functions are the contract; DTO shapes match
    `frontend/src/types/*.ts`.

    - Org scoping comes from the caller's JWT (`ouId`/`ouHandle`) on WSO2 Cloud,
      or the configured namespace on a local OpenChoreo.
    - A web app is a set of deployment tracks; a track (one branch) is one
      OpenChoreo Component, identified by `trackId` (the Component name).
    - Errors use `{code, message}`. `402 QUOTA_EXCEEDED` = the plan's track quota
      is reached (PAS entitlement gate); `501 NOT_SUPPORTED` = a P1/P2 feature;
      `403 PLAN_REQUIRED` = the feature needs a paid plan (HPA, replicas > 1,
      resources above the defaults, environments past the first);
      `429 GIT_RATE_LIMITED` = the GitHub API rate limit is exhausted (public
      GitHub reads are cached for ~60s); `502 UPSTREAM_ERROR` / `504
      UPSTREAM_TIMEOUT` = OpenChoreo or a platform service failed / timed out.
servers:
  - url: /webapp-hosting/api/v1
security:
  - bearerAuth: []
paths:
  /meta:
    get:
      operationId: getMeta
      summary: Target / org info for the console
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meta'
        default:
          $ref: '#/components/responses/Error'
  /plan:
    get:
      operationId: getPlan
      summary: The org's Web App Hosting plan and its feature limits
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        default:
          $ref: '#/components/responses/Error'
  /projects:
    get:
      operationId: listProjects
      summary: List projects (creates the default project on first login)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: createProject
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Error'
  /projects/{projectId}:
    get:
      operationId: getProject
      summary: Get a project
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Error'
  /projects/{projectId}/environments:
    get:
      operationId: listProjectEnvironments
      summary: Pipeline environments in promotion order
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Environment'
        default:
          $ref: '#/components/responses/Error'
  /projects/{projectId}/webapps:
    get:
      operationId: listWebApps
      summary: List web apps
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebApp'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: createWebApp
      summary: Create a web app (default track Component + first build). 402 QUOTA_EXCEEDED when over the plan's track quota.
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebAppInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebApp'
        default:
          $ref: '#/components/responses/Error'
  /projects/{projectId}/webapps/{webAppId}:
    get:
      operationId: getWebApp
      summary: Get a web app
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebApp'
        default:
          $ref: '#/components/responses/Error'
  /git/github/installations:
    get:
      operationId: listGitHubInstallations
      summary: GitHub App installations of the org (empty locally)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationList'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: bindGitHubInstallations
      summary: Exchange the GitHub OAuth code and bind installations (WSO2 Cloud only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindInstallationRequest'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationList'
        default:
          $ref: '#/components/responses/Error'
  /git/github/repos:
    get:
      operationId: listGitHubRepos
      summary: Repositories of an installation (WSO2 Cloud only)
      parameters:
      - name: installationId
        in: query
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryList'
        default:
          $ref: '#/components/responses/Error'
  /git/branches:
    get:
      operationId: listRepoBranches
      summary: Branches of a repository (public URL, or installation + owner/repo)
      parameters:
      - name: repoUrl
        in: query
        required: false
        schema:
          type: string
      - name: installationId
        in: query
        required: false
        schema:
          type: integer
      - name: owner
        in: query
        required: false
        schema:
          type: string
      - name: repo
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/branches:
    get:
      operationId: listWebAppBranches
      summary: Branches of the web app's repository
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks:
    get:
      operationId: listTracks
      summary: List deployment tracks
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeploymentTrack'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: createTrack
      summary: Create a track from a branch (new Component, first build). 402 QUOTA_EXCEEDED when over quota.
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentTrack'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}:
    delete:
      operationId: deleteTrack
      summary: Delete an undeployed, non-default track
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No content
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/deletable:
    get:
      operationId: checkTrackDeletable
      summary: Whether a track can be deleted
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackDeletableResult'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/auto-deploy:
    put:
      operationId: setTrackAutoDeploy
      summary: Toggle auto-deploy to the first environment after successful builds
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoDeployInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentTrack'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/builds:
    get:
      operationId: listBuilds
      summary: List builds, newest first (steps without logs)
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuildRun'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: triggerBuild
      summary: Trigger a build of the branch head (or a given sha)
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerBuildInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildRun'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/builds/{buildId}/logs:
    get:
      operationId: getBuildLogs
      summary: Build steps with logs (live, else archived)
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      - name: buildId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildRun'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/build-config:
    get:
      operationId: getBuildConfig
      summary: Build settings
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildConfig'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/latest-commit:
    get:
      operationId: getLatestCommit
      summary: Head commit of the track's branch
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatestCommit'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/environments:
    get:
      operationId: listTrackEnvironments
      summary: Per-environment deployment summary
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnvironmentDeployment'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/deployments:
    get:
      operationId: listDeployments
      summary: Current deployment per environment
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: deployBuild
      summary: Deploy a build to an environment
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployBuildInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/deployments/promote:
    post:
      operationId: promoteDeployment
      summary: Promote the release running in one environment to another
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoteInput'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        default:
          $ref: '#/components/responses/Error'
  /webapps/{webAppId}/tracks/{trackId}/environments/{env}/redeploy:
    post:
      operationId: redeployEnvironment
      summary: Reactivate a stopped deployment or rolling-restart a running one
      parameters:
      - name: webAppId
        in: path
        required: true
        schema:
          type: string
      - name: trackId
        in: path
        required: true
        schema:
          type: string
      - name: env
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Meta:
      type: object
      properties:
        target: {type: string, enum: [wso2cloud, openchoreo]}
        gitHubApp: {type: boolean}
        billingEnabled: {type: boolean}
        orgHandle: {type: string}
        namespace: {type: string}
...
    TriggerBuildInput:
      type: object
      properties:
        sha: {type: string}
...